journal posting and pushing into user journals array
